### 2019-01-29 #1219 @danielskatz
@seabbs - once you've made changes (as in
https://github.com/seabbs/getTBinR/pull/52), type ```@whedon generate pdf``` as a new comment here to retry the compilation
