+++ inst/paper/paper.md
@@ -13,7 +13,7 @@ authors:
     orcid: 0000-0001-8057-8037
     affiliation: "1"
 affiliations:
-  - name:  Bristol Medical School: Population Health, University of Bristol, Bristol, UK
+  - name:  'Bristol Medical School: Population Health, University of Bristol, Bristol, UK'
     index: 1
 date: 29 January 2019
 bibliography: paper.bib
