### 2019-01-31 dbbe7b6
added doc badge and update docs
