+++ binder/install.R
@@ -1,0 +1,2 @@
+install.packages("remotes")
+remotes::install_github("seabbs/getTBinR", dependencies = TRUE)
