+++ binder/install.R
@@ -1,2 +1,3 @@
 install.packages("remotes")
+## Install package and all deps from GitHub
 remotes::install_github("seabbs/getTBinR", dependencies = TRUE)
