### 2019-01-31 382a171
updated binder to rstudio
