### 2019-01-31 4492f39
moved to dockerfile for binder
