+++ binder/install.R
@@ -1,3 +1,0 @@
-install.packages("remotes")
-## Install package and all deps from GitHub
-remotes::install_github("seabbs/getTBinR", dependencies = TRUE)
