### 2019-01-31 e2cb13c
added magick to dockerfile
