+++ binder/Dockerfile
@@ -6,4 +6,14 @@ RUN chown -R ${NB_USER} ${HOME}
 
 ## Become normal user again
 USER ${NB_USER}
+
+RUN apt-get update -y && \
+    apt-get install -y \
+    texlive-latex-recommended \
+    texlive-fonts-extra \
+    texinfo \
+    libqpdf-dev \
+    libmagick++-dev \
+    && apt-get clean
+    
 RUN R -e "devtools::install_github('seabbs/getTBinR', dependencies = TRUE);"
