+++ binder/Dockerfile
@@ -4,9 +4,6 @@ USER root
 COPY . ${HOME}
 RUN chown -R ${NB_USER} ${HOME}
 
-## Become normal user again
-USER ${NB_USER}
-
 RUN apt-get update -y && \
     apt-get install -y \
     texlive-latex-recommended \
@@ -15,5 +12,8 @@ RUN apt-get update -y && \
     libqpdf-dev \
     libmagick++-dev \
     && apt-get clean
+    
+## Become normal user again
+USER ${NB_USER}
     
 RUN R -e "devtools::install_github('seabbs/getTBinR', dependencies = TRUE);"
