+++ binder/Dockerfile
@@ -7,4 +7,4 @@ RUN chown -R ${NB_USER} ${HOME}
 ## Become normal user again
 USER ${NB_USER}
     
-RUN R -e "devtools::install_dev_deps(); devtools::install()"
+RUN Rscript -e "devtools::install_dev_deps(); devtools::install()"
