### 2019-01-31 26dbe0a
moved back to R from Rscript as causing binder to hang
