### 2019-02-11 #1219 @seabbs
Hi @trallard  - no problem.

No external people in mind. Looking at the list (from bottom to top) here are a few that looked relevant (in no particular order):

* stulacy (R + epidemiology)
* rrrlw (R + Data visualisation + Epidemiology)
* strengejacke (R + Epidemiology + Data visualisation)
* aj2duncan (R + Epidemiology)
* patatefritte (R + Epidemiology)
* ellessenne (R + Epidemiology)
* ibontama (R + Epidemiology)
* rosap (R + Epidemiology
* simonthelwall (R + Epidemiology + Infectious disease)*
* darokun (R + Epidemiology + Public Health)*
* petermacp (R + Epidemiology + Infectious disease)*
* noamross (R + Epidemiology)
* masalmon (R + Epidemiology) *

`*` indicate people that look like they have R experience, are generally active in epidemiology, and have done work on infectious disease.

Hope that helps - let me know if you need anything more.
