### 2019-02-15 #1219 @seabbs
@trallard

I was wondering if It was okay to do development work whilst awaiting/under review? I have the next version [changes prepped](https://github.com/seabbs/getTBinR/projects/2) and just waiting on this to implement (some time has come up whilst I avoid writing my thesis :)).

Thanks!
