### 2019-02-16 #1219 @trallard
Dear reviewers,

@stulacy Stuart are you able and willing to take this review?
@rrrlw Raul are you able and willing to take this review?
@strengejacke Daniel are you able and willing to take this review?
@noamross Noam are you able and willing to take this review?
@maelle MaëlleMaëlle are you able and willing to take this review?
@darokun Daloha are you able and willing to take this review?

**Title:** Get TB Data in R
**Summary:** Quickly and easily import analysis ready TB burden data, from the World Health Organisation (WHO), into R. The aim of the package is to speed up access to high quality TB burden data, using a simple R interface. Generic plotting functions are provided to allow for rapid graphical exploration of the WHO TB data.
**Article Proof:** [https://github.com/openjournals/joss-papers/blob/joss.01219/joss.01219/10.21105.joss.01219.pdf](https://github.com/openjournals/joss-papers/blob/joss.01219/joss.01219/10.21105.joss.01219.pdf)
Submitting author: @seabbs
**Repository:** [https://github.com/seabbs/getTBinR](https://github.com/seabbs/getTBinR)
