### 2019-02-18 #1260 @whedon
[REVIEW]: getTBinR: an R package for accessing and summarising the World Health Organisation Tuberculosis data
**Submitting author:** @seabbs (<a href="http://orcid.org/0000-0001-8057-8037">Sam Abbott</a>)
**Repository:** <a href="https://github.com/seabbs/getTBinR" target ="_blank">https://github.com/seabbs/getTBinR</a>
**Version:** 0.5.8
