### 2019-02-18 #1260 @trallard
👋 @rrrlw and @strengejacke this is the review issue the latest version of the JOSS paper is just up here ⬆

Please use this issue to conduct your review and discuss any issues/concerns raised during the review process.

If you have any questions at any point, please feel free to ping me
