### 2019-02-18 seabbs/getTBinR#67 @rrrlw
Incomplete LICENSE file
GPL-3 is an appropriate OSI-approved license for a JOSS submission. However, the bottom of the LICENSE file in this repo is unfilled (note the use of angled brackets near the bottom, e.g. <name of author>, <year>).

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/1260).
