### 2019-02-18 #1260 @rrrlw
Minor point: the GitHub README meets the criteria for "Statement of need" in the Documentation section (corresponding box has been checked), however, a slightly more detailed description in the DESCRIPTION file could be helpful for useRs who download the CRAN package w/o looking at the GitHub page.
