### 2019-02-18 #1260 @rrrlw
Minor point: the Depends, Imports, and Suggests fields in the DESCRIPTION file meet the criteria for "Installation instructions" (corresponding box has been checked), but it is often helpful to include minimum package versions for the Imports and Suggests fields (as you do for the R dependency under the Depends field).
