### 2019-02-18 #1260 @rrrlw
Quick update: this is an easy submission to review: wonderful & useful software, evidenced by numerous examples, vignettes, sample data, smooth installation, clean pkgdown [website](https://www.samabbott.co.uk/getTBinR/); overall this looks like a mature package with a good number of users that will be useful to researchers.

There are a couple of minor points (as stated above + minor LICENSE point opened as [issue](https://github.com/seabbs/getTBinR/issues/67) in repo) that could improve the package (at the author's discretion); I will spend some more time taking a more detailed look at the vignettes and package code to confirm the function claims of the software are met. Barring any major surprises, I foresee the last couple of boxes being checked soon and my review being complete.

Great work, @seabbs!
