### 2019-02-19 seabbs/getTBinR#67 @seabbs
Hi @rrrlw,

Thanks for taking the time to review this.

Could you give me the line numbers for the areas I need to change?

Thanks!
