### 2019-02-19 seabbs/getTBinR#67 @seabbs
Looked at [here](https://github.com/tidyverse/dplyr/blob/master/LICENSE.md) as an example.

Filled in the <> brackets with relevant details.

Let me know if this needs changing etc.
