### 2019-02-19 #1260 @seabbs
Hi @rrrlw,

Thanks for reviewing.

Do you have a suggestion for how to establish the minimum package versions?

Barring an automated approach, the solutions I see is:

1. Testing the R version given as the minimum and testing the packages available at that time (using rocker docker containers to get package versions).
2. Rolling forward the package versions until all tests pass.

Any other solution appreciated as this seems a little tedious!

Modified the license and description as suggested.

Sam
