### 2019-02-19 3e3e937
minor review changes from @rrrlw - incremented package version
