+++ DESCRIPTION
@@ -5,7 +5,12 @@ Title: Get WHO Tuberculosis Data
 Authors@R: c(person("Sam Abbott", role = c("aut", "cre"),
                      email = "contact@samabbott.co.uk",
                      comment = c(ORCID = "0000-0001-8057-8037")))
-Description: Quickly and easily import analysis ready Tuberculosis (TB) burden data, from the World Health Organisation (WHO), into R. The aim of the package is to speed up access to high quality TB burden data, using a simple R interface. Generic plotting functions are provided to allow for rapid graphical exploration of the WHO TB data. A shiny dashboard is built in to showcase package functionality. An automated report allows country level TB burden to be summarised.
+Description: Quickly and easily import analysis ready Tuberculosis (TB) burden data, 
+    from the World Health Organisation (WHO), into R. The aim of the package is to speed 
+    up access to high quality TB burden data, using a simple R interface. Generic plotting
+    functions are provided to allow for rapid graphical exploration of the WHO TB data. 
+    A shiny dashboard is built in to showcase package functionality. An automated report
+    allows country level TB burden to be summarised.
 License: GPL-3 | file LICENSE
 VignetteBuilder: knitr
 URL: https://www.samabbott.co.uk/getTBinR, 
