### 2019-02-19 dfabdfe
travis update based on that used in dplyr
