+++ .travis.yml
@@ -2,20 +2,37 @@
 
 language: R
 sudo: false
+dist: trusty
 cache: packages
+latex: false
+fortran: false
 
-before_install:
-  - sudo apt-get update
-  - sudo apt-get install -y libmagick++-dev wget
-  
-r:
-  - oldrel
-  - release
-  - devel
+jobs:
+  include:
+  - r: devel
+  - r: 3.1
+    warnings_are_errors: false
+  - os: osx
+  - r: release
+    after_success:
+    - Rscript -e 'covr::codecov()'
+    before_cache:
+    - Rscript -e 'remotes::install_github("r-lib/pkgdown")'
+    deploy:
+      provider: script
+      script: Rscript -e 'pkgdown::deploy_site_github(verbose = TRUE)'
+      skip_cleanup: true
+  - r: 3.4
+  - r: 3.3
+  - r: 3.2
+    warnings_are_errors: false
 
-r_packages:
-  - covr
+matrix:
+  fast_finish: true
 
-after_success:
-  - Rscript -e 'covr::codecov()'
-  - wget https://mybinder.org/v2/gh/seabbs/getTBinR/master?urlpath=rstudio
+env:
+  global:
+  - _R_CHECK_FORCE_SUGGESTS_=false
+  - MAKEFLAGS="-j 2"
+  - TRAVIS_CXXFLAGS="-Wall -Wextra -pedantic -Werror"
+  - R_REMOTES_NO_ERRORS_FROM_WARNINGS=true
