+++ dev/pkgdown.yml
@@ -1,0 +1,11 @@
+pandoc: '2.2'
+pkgdown: 1.3.0.9000
+pkgdown_sha: 7e788c41d20ab28d4bf97556e18382391d359c9f
+articles:
+  case_fatality_rate: case_fatality_rate.html
+  case_study_global_trends: case_study_global_trends.html
+  intro: intro.html
+urls:
+  reference: samabbott.co.uk/getTBinR//reference
+  article: samabbott.co.uk/getTBinR//articles
+
