### 2019-02-19 10910c2
make release docs using travis
