+++ NEWS.md
@@ -7,6 +7,9 @@
 * Added package information to license file - suggested during review for JOSS submission by @[rrrlw](/rrrlw).
 * Improved package DESCRIPTION for CRAN only users - suggested during review for JOSS submission by [rrrlw](/rrrlw).
 * Added development documentation badge to the README + website.
+* Moved to automated pkgdown deployment using travis. Based on [this](https://pkgdown.r-lib.org/reference/deploy_site_github.html) and the [dplyr](https://github.com/tidyverse/dplyr/blob/master/.travis.yml) implementation.
+* Expanded travis testing grid based on [dplyr](https://github.com/tidyverse/dplyr/blob/master/.travis.yml) implementation.
+* Updated earliest supported R version based on travis testing - now `3.1.2`. 
 
 # getTBinR 0.5.7
 
