+++ DESCRIPTION
@@ -20,7 +20,7 @@ Encoding: UTF-8
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
 LazyData: true
-Depends: R (>= 2.10)
+Depends: R (>= 3.1.2)
 Suggests: devtools,
     testthat,
     vdiffr,
