### 2019-02-19 48e74b3
added package versions using usethis::use_tidy_versions, dropped docs
