### 2019-02-19 9a4d3cd
udpated figure paths as lost during commits
