### 2019-02-19 #1260 @seabbs
Hi @rrrlw

Just an update on dep versions. I have updated the R dependency to be `3.1.2` rather than `2.1`. This is the earliest R version that I am actively testing on travis so will definitely work. `2.1` was set by devtools (I think) and I don't really have a guarantee that it works (:().

For package dependencies I have used `usethis::use_tidy_versions()` to set package deps based on what my development environment has loaded.  This looks to be what [dplyr is doing](https://github.com/tidyverse/dplyr/commit/98fafed9566d3fdee1a8f2df26c43493d8f65854). I have also added this to the Makefile so this will update as I do further work.

Let me know if you think this makes sense.

Sam
