### 2019-02-19 seabbs/getTBinR#69 @strengejacke
[JOSS] Statement of Need
Minor issue:
This issue is related to https://github.com/openjournals/joss-reviews/issues/1260

Please add a sentence to the readme that emphasizes the need of this package (e.g., there's no other package, or there's no straightforward way to get the data w/o getTBinR, ...).
