### 2019-02-19 seabbs/getTBinR#69 @seabbs
Hi @strengejacke

Thanks for reviewing this.

The README currently has: "Quickly and easily import analysis ready TB burden data, from the World Health Organisation (WHO), into R. The aim of the package is to speed up access to high-quality TB burden data, using a simple R interface. Generic plotting functions are provided to allow for rapid graphical exploration of the WHO TB data. A shiny dashboard is built in to showcase package functionality. An automated report allows country level TB burden to be summarised."

Not sure what to add here to close this issue. There is no other package but then I would hardly expect there to be for this use case - feels a little odd to explicitly point it out?

The data can be fairly straightforwardly imported by downloading, reading into R (via readr::read_csv for example) and the munging (joining + cleaning). The use case for this package is really to speed this up, make using the dataset easier (i.e interacting with the data dictionary) and embed these data into visualisations etc to make the data more accessible/ analysis more reproducible.  Personally, I would say this is stated in the above?

Very happy to take suggestions 😄

Thanks,

Sam
