### 2019-02-19 e52b840
"added makefile commit and update docs to dev"
