+++ _pkgdown.yml
@@ -7,7 +7,7 @@ template:
       index_name: gettbinr
 
 development:
-  mode: release
+  mode: devel
   
 navbar:
   type: inverse
