### 2019-02-19 seabbs/getTBinR#69 @strengejacke
Well, it's actually not much. The readme already says

> The aim of the package is to speed up access to high-quality TB burden data, using a simple R interface

I think you need just a small sentence, probably like in the paper where you describe the target audience ("non-specialists"). So, for whom is this package for? It's really just a short addition, in order to make the paragraph in the readme look more "complete" - you of course also add a couple of sentences, if the wording looks better for you then ;-) but I don't expect you to write a complete paper here.
