### 2019-02-19 #1260 @rrrlw
The R dependency is always the trickiest for me as well because updating R can be a pain for many useRs; I think you have selected a reasonable minimum (although I think 2.1 was also acceptable). Given that installing/updating a package is generally not as troublesome, your solution of using the versions in your environment is ideal. Thank you for the rapid changes.
