### 2019-02-19 916f8cd
reworked package intro in readme and description
