### 2019-02-19 fc37e5a
improved dockerfile commenting
