+++ Dockerfile
@@ -14,12 +14,14 @@ RUN apt-get update -y && \
 
 ADD . /home/rstudio/getTBinR
 
+## Install pkgdown for website generation
 RUN Rscript -e 'devtools::install_github("r-lib/pkgdown")'
 
 ## Install hexsticker to generate package badge.
 RUN Rscript -e 'install.packages("hexSticker")'
 
-
+## Install dev deps
 RUN Rscript -e 'devtools::install_dev_deps("home/rstudio/getTBinR")'
 
+## Install the getTBinR
 RUN Rscript -e 'devtools::install("home/rstudio/getTBinR")'
