### 2019-02-19 seabbs/getTBinR#69 @seabbs
Good point!

I have added a pull request (#70) with some changes to the README based on the draft JOSS paper (+ updated the package description). Good to look back at the README intro as it hasn't been updated since the first package release.

Let me know if you have suggestions for improving further.

Thanks,

Sam
