### 2019-02-19 857f571
updated news with changes
