+++ NEWS.md
@@ -5,7 +5,9 @@
 ## Package updates
 
 * Added package information to license file - suggested during review for JOSS submission by @[rrrlw](/rrrlw).
-* Improved package DESCRIPTION for CRAN only users - suggested during review for JOSS submission by [rrrlw](/rrrlw).
+* Updated README introduction to better explain package aim -  suggested during review for JOSS submission by @[strengejacke](/strengejacke)
+* Improved package DESCRIPTION for CRAN only users - suggested during review for JOSS submission by @[rrrlw](/rrrlw).
+* Used `usethis::use_tidy_description` to improve DESCRIPTION formatting.
 * Added development documentation badge to the README + website.
 * Moved to automated pkgdown deployment using travis. Based on [this](https://pkgdown.r-lib.org/reference/deploy_site_github.html) and the [dplyr](https://github.com/tidyverse/dplyr/blob/master/.travis.yml) implementation.
 * Expanded travis testing grid based on [dplyr](https://github.com/tidyverse/dplyr/blob/master/.travis.yml) implementation.
