### 2019-02-19 #1260 @strengejacke
As far as I'm concerned and what I have tested and seen from the package, all remaining issues have been addressed by @seabbs. The package documentation, examples and descriptions are pretty straightforward and clear. The same holds true for the paper. Thus, I recommend this paper for publication in JOSS
