### 2019-02-19 a263ca8
added fortran to travis for igraph and hexbin
