+++ DESCRIPTION
@@ -26,7 +26,7 @@ URL: https://www.samabbott.co.uk/getTBinR,
     https://github.com/seabbs/getTBinR
 BugReports: https://github.com/seabbs/getTBinR/issues
 Depends: 
-    R (>= 3.2.1)
+    R (>= 3.2.0)
 Imports: 
     data.table (>= 1.12.0),
     dplyr (>= 0.7.8),
