+++ .travis.yml
@@ -5,7 +5,7 @@ sudo: false
 dist: trusty
 cache: packages
 latex: false
-fortran: false
+fortran: TRUE
 
 jobs:
   include:
