### 2019-02-19 764806d
added fortran to travis for igraph and hexbin
