+++ .travis.yml
@@ -5,7 +5,7 @@ sudo: false
 dist: trusty
 cache: packages
 latex: false
-fortran: TRUE
+fortran: true
 
 jobs:
   include:
