### 2019-02-19 8b4f12c
updated news with 3.2.0 as earliest R version
