### 2019-02-20 #1260 @seabbs
Thank you both (@rrrlw and @strengejacke) for taking the time to review.

Very helpful feedback from you both and a very smooth process.
