### 2019-02-20 f289c04
updated package intro and title + skeleton of citation
