### 2019-02-26 #1260 @seabbs
Will do.

Is this a GitHub release or CRAN release?
