### 2019-02-26 #1260 @trallard
We normally ask for the GitHub version as this is also the one that will be used in Zenodo, but thinking about it: it would be great to have all versions in sync (GitHub, CRAN, Zenodo) of course if you have the time/capacity to do so
