### 2019-02-26 #1260 @seabbs
I'd rather wait on the CRAN release if possible as my last CRAN release was only a month ago. I also have some features planned that I would like to implement in the next release. The issues above were fairly cosmetic (but very useful) so package functionality is unchanged between CRAN and GitHub. If you think the benefits from syncing up outweigh above then happy to release today.

In terms of the Zenodo DOI is it the one for the specific release or is it the one that links to the latest release?
