### 2019-02-26 #1260 @strengejacke
Usually, you create a new Zenodo-DOI on the latest GitHub version, i.e. the version _after_ review. That's the DOI for the JOSS article.
