### 2019-02-26 #1260 @seabbs
Sorry if I am being thick.

To clarify a DOI that points to the specific version of the package that was reviewed here - rather than a DOI that points to the latest Zenodo release?
