### 2019-02-26 #1260 @strengejacke
You have made some revisions, so you should change your package version number and create a new DOI, so the Zenodo-archive contains all latest changes after review.
