### 2019-02-26 #1260 @trallard
Hey @seabbs no worries this can be a bit confusing let me go over what you'd need to do:
1. Create a new release (tag) on GitHub of the package
2. Create a new Zenodo-DOI pointing to the newly created release (if you already have the code archived in Zenodo it should be a matter of adding the new version)
3. Copy the newly created DOI to this thread
