### 2019-02-26 #1260 @seabbs
Yes I totally agree on that - making a release now.

My point is that Zenodo has two DOI types. One points to a specific release and one tracks the latest release.

From what you are both saying the DOI I need is the one that links to the reviewed package. Obviously, this is good as it shows what was reviewed. The downside is that it will be out of date as soon as there is a new release.
