### 2019-02-26 #1260 @trallard
Sorry I just corrected my previous comment: share the DOI that points to the latest release (that should include the reviewed version)
