+++ inst/paper/paper.bib
@@ -9,12 +9,12 @@
 
 @misc{Abbott:2019,
   author       = {Sam {Abbott}},
-  title        = {seabbs/getTBinR 0.5.7},
+  title        = {seabbs/getTBinR 0.5.8},
   month        = jan,
   year         = {2019},
-  doi          = {10.5281/zenodo.2547406},
+  doi          = {10.5281/zenodo.2547405},
   journal      = {Zenodo},
-  url          = {https://doi.org/10.5281/zenodo.2547406}
+  url          = {https://doi.org/10.5281/zenodo.2547405}
 }
 
 @misc{WHO:2018,
@@ -22,7 +22,7 @@
   title        = {Global Tuberculosis Report},
   month        = oct,
   year         = {2018},
-  url          = { https://www.who.int/tb/country/data/download/en/}
+  url          = {https://www.who.int/tb/country/data/download/en/}
 }
 
 @misc{Abbott:2018,
