### 2019-02-26 #1260 @seabbs
Checklist:

- [x]  make a new release of the package (unless the latest release is up to date with the recent modifications) and add let us know what is the most up to date version
- [x] upload the revised software to your DOI-granting data/software repository, and post the DOI here
- [x] double check the submission paper to ensure it is up-to-date (citations are correct, DOIs are added, etc)

Changes:

* Updated paper with correct latest release package DOI rather than version DOI
* Added Sciences to department naming in paper.
* Updated package version to `0.5.8` and released on GitHub
* Uploaded to Zenodo (via GitHub)
* DOI (to latest release): 10.5281/zenodo.2547405
