### 2019-02-26 #1260 @trallard
@seabbs  all looks good to me so I am going to proceed for acceptance

@rrrlw and @strengejacke thanks a lot for your time and valuable contribution to JOSS as reviewers for this submission 🙌🏻

@openjournals/joss-eics this submission has been accepted and is ready to be published 🎉🙌✨
