### 2019-02-26 #1260 @seabbs
Thanks, @trallard for overseeing this.

Thanks again to @rrrlw  and @strengejacke for the great review.

Great painless experience. It would be great if traditional journals moved towards this kind of model!
