### 2019-02-26 #1260 @danielskatz
Thanks all!  Processing for acceptance will happen later today/tonight.
